+++ NEWS.md
@@ -1,6 +1,11 @@
 # mlr3tuning (development version)
 
-fix: The `as_data_table()` functions do not unnest the `x_domain` colum anymore by default.
+* fix: The `as_data_table()` functions do not unnest the `x_domain` colum anymore by default.
+* fix: `to_tune(internal = TRUE)` now also works if non-internal tuning parameters require have
+  an `.extra_trafo`
+* feat: It is now possible to pass an `internal_search_space` manually.
+  This allows to use parameter transformations on the primary search space in combination with
+  internal hyperparameter tuning.
 
 # mlr3tuning 1.0.2